+++ src/lib.rs
@@ -5,7 +5,7 @@
  */
 
 pub mod compare;
-pub mod config;
+// pub mod config;
 pub mod download;
 pub mod flex_query;
 pub mod flex_reader;